+++ src/replace.js
@@ -1,4 +1,5 @@
 const vscode = require('vscode');
+let lastPickedItem = '';
 const OutputChannel = vscode.window.createOutputChannel(`Output Channel`);
 module.exports = {
     replaceRulesInAllDocuments: function () {
@@ -11,11 +12,14 @@ module.exports = {
     replaceRulesInCurrDocSelection: function () {
         pickAndExcuteRulesetICurrDocSelection();
     },
-    getNewText: function (originalText, searchExpresion, replaceExpression, jsModuleFilePath, jsFunctionName,document,range) {
-        return getNewText(originalText, searchExpresion, replaceExpression, jsModuleFilePath, jsFunctionName,document,range)
+    getNewText: function (originalText, searchExpresion, replaceExpression, jsModuleFilePath, jsFunctionName, document, range) {
+        return getNewText(originalText, searchExpresion, replaceExpression, jsModuleFilePath, jsFunctionName, document, range)
     },
     emptySearchexpressionError: function (searchExpresion, ruleName) {
         return emptySearchexpressionError(searchExpresion, ruleName);
+    },
+    placeLastSelectionInTop: function (ruleSetNames, lastPickedItem) {
+        placeLastSelectionInTop(ruleSetNames, lastPickedItem);
     }
 }
 
@@ -40,9 +44,9 @@ async function replaceRulesInAllDocuments(rules, fileExtension, ruleSetName = ''
                 progress.report({ message: 'Replacing rules in document ' + document.fileName });
                 for (let i = 0; i < rules.length; i++) {
                     let customRule = rules[i];
-                    for (let index = 0; index < 2; index++) {                        
-                        await replaceRuleInDocument(customRule, document, ruleSetName);   
-                    }                    
+                    for (let index = 0; index < 2; index++) {
+                        await replaceRuleInDocument(customRule, document, ruleSetName);
+                    }
                 }
             }
             catch (error) {
@@ -60,7 +64,7 @@ async function replaceRuleInRange(replaceRule, document, replaceRange = new vsco
     for (let index = 0; index < getNumberOfRepetitions(replaceRule); index++) {
         await replaceRuleInRangeOnce(replaceRule, document, replaceRange, ruleSetName);
     }
-    
+
 }
 async function replaceRuleInRangeOnce(replaceRule, document, replaceRange = new vscode.Range(0, 0, 0, 0), ruleSetName = '') {
     if (!replaceRule) {
@@ -69,8 +73,8 @@ async function replaceRuleInRangeOnce(replaceRule, document, replaceRange = new 
     if (emptySearchexpressionError(replaceRule.searchExpresion, replaceRule.name)) {
         return;
     }
-    try{
-    const regex = new RegExp(replaceRule.searchExpresion, 'mgi');
+    try {
+        const regex = new RegExp(replaceRule.searchExpresion, 'mgi');
     }
     catch (error) {
         showErrorRegExp(replaceRule.name, error);
@@ -81,9 +85,9 @@ async function replaceRuleInRangeOnce(replaceRule, document, replaceRange = new 
         return;
     }
     let originalText = document.getText(replaceRange);
-    
+
     const replaceText = getNewText(originalText, replaceRule.searchExpresion, replaceRule.replaceExpression,
-        replaceRule.jsModuleFilePath, replaceRule.jsFunctionName,document,replaceRange);
+        replaceRule.jsModuleFilePath, replaceRule.jsFunctionName, document, replaceRange);
 
     if (replaceText === originalText) {
         return;
@@ -98,10 +102,12 @@ async function replaceRuleInRangeOnce(replaceRule, document, replaceRange = new 
 }
 function pickAndExcuteRuleset() {
     const getRules = require('./getRules.js');
-    const ruleSetNames = getRuleSetNames('','workspace');    
-    //show vscode code picker with the ruleSetNames        
+    const ruleSetNames = getRuleSetNames('', 'workspace');
+    //show vscode code picker with the ruleSetNames
+    placeLastSelectionInTop(ruleSetNames, lastPickedItem);
     vscode.window.showQuickPick(ruleSetNames).then(async (value) => {
         if (value) {
+            lastPickedItem = value;
             let rules = getRules.getRulesFromRuleSetName(value);
             const fileExtension = getRules.getFileExtensionFormRuleSetName(value);
             await replaceRulesInAllDocuments(rules, fileExtension, value);
@@ -109,11 +115,11 @@ function pickAndExcuteRuleset() {
     }
     );
 }
-function getNewText(originalText, searchExpresion, replaceExpression, jsModuleFilePath, jsFunctionName,document,range = new vscode.Range(0, 0, 0, 0)) {
+function getNewText(originalText, searchExpresion, replaceExpression, jsModuleFilePath, jsFunctionName, document, range = new vscode.Range(0, 0, 0, 0)) {
     const existsReplaceExpr = replaceExpression !== undefined;
     let newText = originalText;
     const regex = new RegExp(searchExpresion, 'mgi');
-    
+
     if (jsModuleFilePath) {
         if (jsFunctionName) {
             try {
@@ -122,15 +128,13 @@ function getNewText(originalText, searchExpresion, replaceExpression, jsModuleFi
                 fn = jsModule[jsFunctionName];
                 var setDocumentAndRange = new Function();
                 setDocumentAndRange = jsModule['setDocumentAndRange'];
-                if (setDocumentAndRange)
-                {
-                    setDocumentAndRange(document,range);
+                if (setDocumentAndRange) {
+                    setDocumentAndRange(document, range);
                 }
                 newText = originalText.replace(regex, fn);
                 var getDocument = new Function();
                 getDocument = jsModule['getDocument'];
-                if (getDocument)
-                {
+                if (getDocument) {
                     document = getDocument();
                 }
             }
@@ -157,23 +161,34 @@ function emptySearchexpressionError(searchExpression, ruleName) {
 async function pickAndExcuteRulesetICurrDoc() {
     const getRules = require('./getRules.js');
     const currFileExtension = getFileExtensionFromFileName(await vscode.window.activeTextEditor.document.fileName);
-    let ruleSetNames = getRuleSetNames(currFileExtension,'document');
+    let ruleSetNames = getRuleSetNames(currFileExtension, 'document');
     //show vscode code picker with the ruleSetNames        
+    placeLastSelectionInTop(ruleSetNames, lastPickedItem);
     vscode.window.showQuickPick(ruleSetNames).then(async (value) => {
         if (value) {
+            lastPickedItem = value;
             let rules = getRules.getRulesFromRuleSetName(value);
             await replaceRulesInCurrDoc(rules);
         }
     }
     );
 }
+function placeLastSelectionInTop(ruleSetNames, lastPickedItem) {
+    const indexToRemove = ruleSetNames.indexOf(lastPickedItem);
+    if (indexToRemove >= 0) {
+        ruleSetNames.splice(indexToRemove, 1);
+        ruleSetNames.unshift(lastPickedItem);
+    }
+}
 async function pickAndExcuteRulesetICurrDocSelection() {
     const getRules = require('./getRules.js');
     const currFileExtension = getFileExtensionFromFileName(await vscode.window.activeTextEditor.document.fileName);
-    let ruleSetNames = getRuleSetNames(currFileExtension,'selection');
+    let ruleSetNames = getRuleSetNames(currFileExtension, 'selection');
     //show vscode code picker with the ruleSetNames        
+    placeLastSelectionInTop(ruleSetNames, lastPickedItem);
     vscode.window.showQuickPick(ruleSetNames).then(async (value) => {
         if (value) {
+            lastPickedItem = value;
             let rules = getRules.getRulesFromRuleSetName(value);
             await replaceRulesInCurrDocSelection(rules);
         }
@@ -204,7 +219,7 @@ async function replaceRulesInCurrDocSelection(rules) {
         await replaceRuleInRange(customRule, document, selectioRange, '');
     }
 }
-function getRuleSetNames(fileExtension = '',scope='') {
+function getRuleSetNames(fileExtension = '', scope = '') {
     const getRules = require('./getRules.js');
     let ruleSets = getRules.getRuleSets();
     if (fileExtension !== '') {
@@ -215,29 +230,24 @@ function getRuleSetNames(fileExtension = '',scope='') {
     }
     let ruleSetNames = [];
     for (let i = 0; i < ruleSets.length; i++) {
-        if (GetApplyScope(ruleSets[i],scope))
-        {
+        if (GetApplyScope(ruleSets[i], scope)) {
             ruleSetNames.push(ruleSets[i].name);
-        }        
+        }
     }
     return ruleSetNames;
 }
 function filterRuleSets(ruleSets = {}, currFileExtension = '') {
-    return ruleSets.filter(ruleSetName => ruleSetName.fileExtension === currFileExtension);
+    return ruleSets.filter(ruleSetName => ruleSetName.fileExtension === currFileExtension || ruleSetName.fileExtension === '');
 }
-function GetApplyScope(ruleSet={},scope='')
-{
-    if (!ruleSet.scope)
-    {
+function GetApplyScope(ruleSet = {}, scope = '') {
+    if (!ruleSet.scope) {
         return true;
     }
-    if (ruleSet.scope.length == 0)
-    {
+    if (ruleSet.scope.length == 0) {
         return true;
     }
     for (let index = 0; index < ruleSet.scope.length; index++) {
-        if (ruleSet.scope[index] == scope)
-        {
+        if (ruleSet.scope[index] == scope) {
             return true;
         }
     }
@@ -266,11 +276,9 @@ function showErrorRegExp(ruleName = '', errorRaised) {
     vscode.window.showErrorMessage(finalMessage);
     OutputChannel.appendLine(finalMessage);
 }
-function getNumberOfRepetitions(replaceRule)
-{
-    if (replaceRule.numberOfRepetitions)
-    {
-        return replaceRule.numberOfRepetitions ;
+function getNumberOfRepetitions(replaceRule) {
+    if (replaceRule.numberOfRepetitions) {
+        return replaceRule.numberOfRepetitions;
     }
     return 1;
 }